requirements.txt
fastapi
uvicorn
requests
python-dotenv
httpx
firebase-admin
